requirements.txt
Flask==3.1.0
requests==2.32.3
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-system-metrics
